NAVI/feat/authorised,notauthorised/layout.tsx
two albums page:

1) album = main album page
2) album = in album artist page

(after backend is ready, album artist pages are connected on main )
