+++ app/(authorised)/album/page.module.scss
