+++ app/(authorised)/artists/page.module.scss
