+++ app/(authorised)/artists/page.tsx
@@ -1,18 +1,19 @@
 'use client'
 
-import Playlist from "../components/Playlists/Playlist";
+
+import { ArtistsPage } from "@/app/components/Artists/Artists";
 import styles from "./page.module.scss";
 import { useEffect } from "react";
 
-export default function PlaylistPage() {
+export default function Artists() {
 
   useEffect(() => {
-    document.title= 'Chakrulos | Playlist '
+    document.title= 'Chakrulos | Artists '
   },[]);
 
   return (
     <main className={styles.main}>
-        <Playlist />
+        <ArtistsPage />
     </main>
   );
 }
