+++ app/(authorised)/favorites/page.module.scss
