+++ app/(authorised)/playlist/page.tsx
@@ -1,18 +1,18 @@
 'use client'
 
-import { ArtistsPage } from "../components/Artists/Artists";
+import Playlist from "@/app/components/Playlists/Playlist";
 import styles from "./page.module.scss";
 import { useEffect } from "react";
 
-export default function Artists() {
+export default function PlaylistPage() {
 
   useEffect(() => {
-    document.title= 'Chakrulos | Artists '
+    document.title= 'Chakrulos | Playlist '
   },[]);
 
   return (
     <main className={styles.main}>
-        <ArtistsPage />
+        <Playlist />
     </main>
   );
 }
