+++ app/(notauthorised)/Signup/page.module.scss
@@ -1,4 +1,4 @@
-@import "../Colors/Color.scss";
+@import "../../Colors/Color.scss";
 
 .main {
   display: flex;
