+++ app/(notauthorised)/layout.tsx
@@ -1,0 +1,11 @@
+export default function RootLayout({
+    children,
+  }: Readonly<{
+    children: React.ReactNode;
+  }>) {
+    return (
+      <div>
+          {children}
+      </div>
+    );
+  }
