+++ app/components/Albums/Albums.tsx
@@ -4,7 +4,7 @@ import styles from "./Albums.module.scss"
 import Link from "next/link";
 import AlbumCard from "./AlbumCard/AlbumCard";
 import { Search } from "../Header/Search/Search";
-import { albumsState } from "@/app/states";
+import { albumsState } from "@/app/(authorised)/states";
